prevent removal if using travisci

+++ hooks/post_gen_project.py
@@ -69,8 +69,11 @@ def remove_utility_files():
 
 
 def remove_heroku_files():
-    file_names = ["Procfile", "runtime.txt"]
+    file_names = ["Procfile", "runtime.txt", "requirements.txt"]
     for file_name in file_names:
+        if file_name == "requirements.txt" and "{{ cookiecutter.use_travisci }}".lower() == "y":
+            # don't remove the file if we are using travisci but not using heroku
+            continue
         os.remove(file_name)
